% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_test.R
\name{cor_test}
\alias{cor_test}
\title{Correlation tests}
\usage{
cor_test(var1, var2, round = TRUE, ...)
}
\arguments{
\item{var1}{a data.frame column to compare with \code{var2}}

\item{var2}{a data.frame column to compare with \code{var1}}

\item{round}{logical, whether or not to round the results}
}
\value{
Returns a data.frame version of the standard htest output.
}
\description{
A wrapper for \code{\link[stats]{cor.test}} that returns htest data in a
data.frame rather than a list
}
\examples{
cor_test(rnorm(1:100, 100), rnorm(1:100, 100), method = "pearson")

}

