% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htest_data_frame.R
\name{dfh}
\alias{dfh}
\title{Convert an htest list, or a list of htest lists to a data.frame}
\usage{
dfh(htest.list)
}
\arguments{
\item{htest.list}{a single htest list or a list of htest lists}
}
\value{
Returns a \code{data.frame} with data from one or more htests
}
\description{
Convert an htest list, or a list of htest lists to a data.frame
}
\examples{
data(htests)
htest_data_frame(htests[["cor.test.pearson"]])

}

