% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htest_data_frame.R
\name{htest_data_frame}
\alias{htest_data_frame}
\title{Title}
\usage{
htest_data_frame(htest.list)
}
\arguments{
\item{htest.list}{a single htest list or a list of htest lists}
}
\value{
Returns a \code{data.frame} with data from one or more htests
}
\description{
Title
}
\examples{
a <- list()
a[[1]] <- cor.test(rnorm(10), rnorm(10))
a[[2]] <- cor.test(rnorm(10), rnorm(10))
a[[3]] <- cor.test(rnorm(10), rnorm(10))
htest_data_frame(a)

}

