% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_lengths.R
\name{list_lengths}
\alias{list_lengths}
\title{Obtain list of list lengths}
\usage{
list_lengths(data1, overOne = TRUE)
}
\arguments{
\item{data1}{a list of lists}

\item{overOne}{logical, \code{TRUE} returns index of values over one}
}
\value{
A list showing the lengths of the nested list items and the index for
  idenitfying items with a length greater than one.
}
\description{
Summarize the length of all items in a nested list with the the option of
identifying lengths over one.
}
\examples{
exlist1 <- list(list(1, 2, c(3, 0)),
                list(1, 2, c(3, 0)),
                list(1, 2, c(3, 0))
                )
list_lengths(exlist1)

exlist2 <- list(list(1, 2, 3),
                list(1, 2, c(3, 0)),
                list(1, 2, c(3, 0))
)
list_lengths(exlist2)

exlist3 <- list(list(1, c(3, 0), c(3, 0)),
                list(1, c(3, 0), c(3, 0)),
                list(1, c(3, 0), c(3, 0))
)
list_lengths(exlist3)
}

