% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{Student's t-Test}
\usage{
t_test(var1, var2, round = TRUE, ...)
}
\arguments{
\item{var1}{a data.frame column to compare with \code{var2}}

\item{var2}{a data.frame column to compare with \code{var1}}

\item{round}{logical, whether or not to round the results}
}
\value{
Returns a data.frame version of the standard htest output. See
  \code{t.test} for additional arguments and further detail
}
\description{
A wrapper for \code{\link[stats]{t.test}} that returns htest data in a
data.frame rather than a list
}
\examples{
t_test(rnorm(10, 5), rnorm(10, 5))

}

