% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_test.R
\name{wilcox_test}
\alias{wilcox_test}
\title{Wilcoxon Rank Sum and Signed Rank Tests}
\usage{
wilcox_test(var1, var2, round = TRUE, ...)
}
\arguments{
\item{var1}{a vector to correlate with var2}

\item{var2}{a vector to correlate with var1}

\item{round}{logical, whether or not to round the results}

\item{method}{the correlation method to use, either pearson, spearman, or
kendall}
}
\value{
Returns a data.frame version of the standard htest output. See
  \code{ks.test} for additional arguments and further detail
}
\description{
A wrapper for \code{\link[stats]{wilcox.test}} that returns htest data in a
data.frame rather than a list
}
\examples{
wilcox_test(rnorm(10, 5), rnorm(10, 5))

}

