% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htest_data_frame.R
\name{htest_data_frame}
\alias{htest_data_frame}
\title{Convert an htest list, or a list of htest lists to a data.frame}
\usage{
htest_data_frame(data1)
}
\arguments{
\item{htest.list}{a single htest list or a list of htest lists}
}
\value{
Returns a \code{data.frame} where each row represents a list from
one or more htests
}
\description{
To see the list of compatible htest lists run: names(htests)
}
\examples{
data(htests)
htest_data_frame(htests[["cor.test.pearson"]])

}

